Logic - Creation of private room done.
