+++ src/main/java/match/entities/utils/Location.java
@@ -1,0 +1,24 @@
+package match.entities.utils;
+
+public class Location {
+    private double x;
+    private double y;
+
+    public Location(double x, double y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public double getX() {
+        return x;
+    }
+
+    public double getY() {
+        return y;
+    }
+
+    public void setLocation(double x, double y) {
+        this.x = x;
+        this.y = y;
+    }
+}
