--- src/main/java/player/Location.java
@@ -1,24 +1,0 @@
-package player;
-
-public class Location {
-    private double x;
-    private double y;
-
-    public Location(double x, double y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    public double getX() {
-        return x;
-    }
-
-    public double getY() {
-        return y;
-    }
-
-    public void setLocation(double x, double y) {
-        this.x = x;
-        this.y = y;
-    }
-}
