+++ src/main/java/servers/connection/Connection.java
@@ -1,0 +1,41 @@
+package servers.connection;
+
+import utils.GlobalSettings;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.net.SocketTimeoutException;
+
+public abstract class Connection
+{
+    protected boolean m_IsConnected;
+    private long m_LastConnectionTime;
+
+    public Connection()
+    {
+        this.m_IsConnected = true;
+        this.UpdateLastConnectionTime();
+    }
+
+    protected void UpdateLastConnectionTime()
+    {
+        this.m_LastConnectionTime = System.currentTimeMillis();
+    }
+
+    protected boolean isTimedOut()
+    {
+        return (System.currentTimeMillis() - this.m_LastConnectionTime > GlobalSettings.MAX_TIME_OUT);
+    }
+
+    protected boolean ValidateConnectionNeeded()
+    {
+        return (System.currentTimeMillis() - this.m_LastConnectionTime > 5000);
+    }
+
+    abstract public void SendMessage(String i_Message) throws SocketTimeoutException;
+    abstract public void CloseConnection(String i_ExceptionMessage);
+    abstract public boolean IsConnectionAlive();
+    abstract public String ReadMessage() throws IOException;
+    abstract public String getHost();
+    abstract public String WaitForPlayerResponse() throws IOException;
+}
