+++ src/main/java/servers/connection/webconnection/WebConnection.java
@@ -1,0 +1,13 @@
+package servers.connection.webconnection;
+
+
+public final class WebConnection {
+
+
+}
+
+
+
+
+
+
