+++ src/main/java/utils/json/JsonFormatter.java
@@ -3,6 +3,7 @@ package utils.json;
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 
 public class JsonFormatter {
 
@@ -12,6 +13,12 @@ public class JsonFormatter {
         JsonElement jsonElement = gson.fromJson(json, JsonElement.class);
         return jsonElement.getAsJsonObject();
     }
+
+    public JsonElement stringToJsonElement(String jsonString) {
+        // Parsing the string into a JsonElement
+        return GetGson().fromJson(jsonString, JsonElement.class);
+    }
+
 
     public static Gson GetGson() {
         return gson;
